+++ jabgui/src/main/java/org/jabref/gui/keyboard/CodeAreaKeyBindings.java
@@ -134,9 +134,10 @@ public class CodeAreaKeyBindings {
                                                    : NavigationActions.SelectionPolicy.CLEAR;
 
         boolean optionOnly = event.isAltDown() && !event.isMetaDown() && !event.isControlDown();
+        boolean controlOnly = event.isControlDown() && !event.isMetaDown() && !event.isAltDown();
         boolean commandOnly = event.isMetaDown() && !event.isAltDown() && !event.isControlDown();
 
-        if (optionOnly) {
+        if (optionOnly || controlOnly) {
             if (code == KeyCode.LEFT) {
                 codeArea.wordBreaksBackwards(2, policy);
             } else {